Observe the multi_face_landmarks stream
Call ObserveOutputStream to observe the multi_face_landmarks stream, in
an attempt to retrieve the face landmarks when present

Also, call AssertOk() for safety and use for (;;) instead of while (true).

Currently works for a significant amout of time without taking more than
about 600 MB on its own, but aborts due to a stack smash error after
about 1950 frames.

Co-authored-by: homuler <homuler@users.noreply.github.com>